Working Version of Day2 Part1: Cube Game
This is the solution I used to get the correct answer for Part1 Day2. It consists of simply finding the max value of each "color" by using regex to parse the string and find the values corresponding to each color, then checking the max value of that list and checking it against the rules of the game.